+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## 3.2.1 (*/*/*)
+## 3.3.0 (*/*/*)
+  * 'initialState' for context providers while constructing servers before provision
+  * 'queryDataSync' for sync queries execution without promises
   * Signal type update without optional data parameter
   * Default param value for 'asMerged' method for nested objects
   * Better typing/type derivation for Signal/SignalEvent types